requirements.txt
# AI Study Buddy - Python Backend Dependencies

# Web Framework (API)
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6

# PDF Processing
pypdf>=3.17.0

# Word Document Processing
python-docx>=1.1.0

# Vector Database (local, persistent)
chromadb>=0.4.0

# HTTP requests for Ollama API
requests>=2.31.0

# Legacy Streamlit (optional, for fallback UI)
# streamlit>=1.28.0

# Everything runs locally using Ollama - $0 cost!
